fixes based on feedback
